api_build_for_func/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pillow==11.0.0

